pyproject.toml
[project]
name = "cleaning-service"
version = "0.1.0"
description = "web application for cleaning services"
authors = [
    {name = "alexander",email = "alex.matveev.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "async-typer (>=0.1.8,<0.2.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.38,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "orjson (>=3.10.15,<4.0.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "psycopg[binary,pool] (>=3.2.5,<4.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "pytest-postgresql (>=7.0.0,<8.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.15.0"
pre-commit = "^4.1.0"

[tool.mypy]
strict = true

[tool.black]
line-length = 119
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
  | \.env
  | \.flake8
)/
'''

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
split_on_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["fastapi-application"]
skip = [".gitignore", ".dockerignore"]

[tool.flake8]
ignore = ['E231', 'E241']
per-file-ignores = [
    '__init__.py:F401',
]
max-line-length = 119
max-complexity = 8
count = true
extend-immutable-calls = [
    "Query",
    "Depends",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
